Double Fix of Map01
So Map01 got deleted, so we remade it. And then had to fix all the scripts that were changed after the deletion.

+++ COP4331TD/Assets/Scripts/ShootAtEnemy.cs
@@ -23,7 +23,7 @@ public class ShootAtEnemy : MonoBehaviour {
             turret.transform.rotation = Quaternion.LookRotation(newDirection);
         }*/
 
-        if (enemies.Count != 0) {
+        if ((int)enemies.Count != 0) {
             GameObject firstEnemy = (GameObject) enemies.Peek();
             Vector3 targetDirection = firstEnemy.transform.position - turret.transform.position;
             float singleStep = 1000 * Time.deltaTime;
